requirements.txt
pvrecorder
speechrecognition
python-dotenv
